backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.6
pydantic==2.7.1
pydantic-settings==2.2.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
statsmodels==0.14.2
tensorflow==2.16.1
keras==3.3.3
httpx==0.27.0
apscheduler==3.10.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
loguru==0.7.2
matplotlib==3.8.4
scipy==1.13.0
